fix: allow floatable passive numbers

+++ Common/Systems/PassiveTreeSystem/PassiveTreePlayer.cs
@@ -373,7 +373,7 @@ internal class PassiveTreePlayer : ModPlayer
 		return new(false, []);
 	}
 
-	internal void AllocatePassive(Passive passive, int strength = 1, bool save = true)
+	internal void AllocatePassive(Passive passive, float strength = 1, bool save = true)
 	{
 		if (passive is MasteryPassive) // Hardcode for this, which doesn't work the same as any other passive
 		{
@@ -391,7 +391,7 @@ internal class PassiveTreePlayer : ModPlayer
 		}
 	}
 
-	internal void DeallocatePassive(Passive passive, int valueLoss, int pointRefund, bool save = true)
+	internal void DeallocatePassive(Passive passive, float valueLoss, int pointRefund, bool save = true)
 	{
 		if (passive is MasteryPassive)
 		{
